+++ src/redux/actions.js
@@ -1,7 +1,3 @@
-export const ERROR = 'app/error';
 export const RESET = 'app/reset';
 
-export const setError = (payload) => ({
-  type: ERROR,
-  payload,
-});
+export default {};
