--- src/redux/error/index.js
@@ -1,16 +1,0 @@
-import { ERROR } from '../actions';
-
-const initialState = {
-  message: null,
-};
-
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ERROR:
-      return { message: action.payload };
-    default:
-      return state;
-  }
-};
-
-export default reducer;
